python/src/PythonBasicTeaching/LineBot/Chapter_7/requirements.txt
line-bot-sdk
flask
requests
APScheduler
pandas
lxml
